requirements.txt
beautifulsoup4
requests
pillow
boto3